requirements.txt
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.1
cssselect2==0.8.0
dj-database-url==3.0.1
Django==5.2.5
django-widget-tweaks==1.5.0
gunicorn==23.0.0
html5lib==1.1
idna==3.10
lxml==6.0.2
oscrypto==1.3.0
packaging==25.0
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
paypalrestsdk==1.13.3
pillow==11.3.0
psycopg2-binary==2.9.11
pycparser==2.23
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
pyOpenSSL==25.3.0
pypdf==6.1.0
python-bidi==0.6.6
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
reportlab==4.4.4
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==12.5.0
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
webencodings==0.5.1
wheel==0.45.1
whitenoise==6.11.0
xhtml2pdf==0.2.17
